+++ packages/netlify-cms-app/src/backends.js
@@ -1,12 +1,12 @@
-import { registerBackend } from 'netlify-cms-core';
+import { NetlifyCmsCore as CMS } from 'netlify-cms-core';
 import { GitHubBackend } from 'netlify-cms-backend-github';
 import { GitLabBackend } from 'netlify-cms-backend-gitlab';
 import { GitGatewayBackend } from 'netlify-cms-backend-git-gateway';
 import { BitbucketBackend } from 'netlify-cms-backend-bitbucket';
 import { TestBackend } from 'netlify-cms-backend-test';
 
-registerBackend('git-gateway', GitGatewayBackend);
-registerBackend('github', GitHubBackend);
-registerBackend('gitlab', GitLabBackend);
-registerBackend('bitbucket', BitbucketBackend);
-registerBackend('test-repo', TestBackend);
+CMS.registerBackend('git-gateway', GitGatewayBackend);
+CMS.registerBackend('github', GitHubBackend);
+CMS.registerBackend('gitlab', GitLabBackend);
+CMS.registerBackend('bitbucket', BitbucketBackend);
+CMS.registerBackend('test-repo', TestBackend);