+++ packages/netlify-cms-widget-date/src/index.js
@@ -9,4 +9,4 @@ const Widget = (opts = {}) => ({
 });
 
 export const NetlifyCmsWidgetDate = { Widget, controlComponent, previewComponent };
-export { Widget as default, controlComponent, previewComponent };
+export default NetlifyCmsWidgetDate;