+++ packages/netlify-cms-widget-list/src/index.js
@@ -1,4 +1,13 @@
 import controlComponent from './ListControl';
-import { previewComponent } from 'netlify-cms-widget-object';
-export const NetlifyCmsWidgetList = { controlComponent, previewComponent };
-export { controlComponent, previewComponent };
+import NetlifyCmsWidgetObject from 'netlify-cms-widget-object';
+
+const previewComponent = NetlifyCmsWidgetObject.previewComponent;
+const Widget = (opts = {}) => ({
+  name: 'list',
+  controlComponent,
+  previewComponent,
+  ...opts,
+});
+
+export const NetlifyCmsWidgetList = { Widget, controlComponent, previewComponent };
+export default NetlifyCmsWidgetList;