+++ packages/netlify-cms-widget-map/src/index.js
@@ -1,6 +1,13 @@
 import withMapControl from './withMapControl';
 import previewComponent from './MapPreview';
-const controlComponent = withMapControl();
 
-export const NetlifyCmsWidgetMap = { controlComponent, previewComponent };
-export { controlComponent, previewComponent };
+const controlComponent = withMapControl();
+const Widget = (opts = {}) => ({
+  name: 'map',
+  controlComponent,
+  previewComponent,
+  ...opts,
+});
+
+export const NetlifyCmsWidgetMap = { Widget, controlComponent, previewComponent };
+export default NetlifyCmsWidgetMap;