+++ packages/netlify-cms-widget-markdown/src/index.js
@@ -1,4 +1,12 @@
 import controlComponent from './MarkdownControl';
 import previewComponent from './MarkdownPreview';
-export const NetlifyCmsWidgetMarkdown = { controlComponent, previewComponent };
-export { controlComponent, previewComponent };
+
+const Widget = (opts = {}) => ({
+  name: 'markdown',
+  controlComponent,
+  previewComponent,
+  ...opts,
+});
+
+export const NetlifyCmsWidgetMarkdown = { Widget, controlComponent, previewComponent };
+export default NetlifyCmsWidgetMarkdown;