+++ packages/netlify-cms-widget-object/src/index.js
@@ -1,4 +1,12 @@
 import controlComponent from './ObjectControl';
 import previewComponent from './ObjectPreview';
-export const NetlifyCmsWidgetObject = { controlComponent, previewComponent };
-export { controlComponent, previewComponent };
+
+const Widget = (opts = {}) => ({
+  name: 'object',
+  controlComponent,
+  previewComponent,
+  ...opts,
+});
+
+export const NetlifyCmsWidgetObject = { Widget, controlComponent, previewComponent };
+export default NetlifyCmsWidgetObject;