+++ packages/netlify-cms-widget-select/src/index.js
@@ -1,4 +1,12 @@
 import controlComponent from './SelectControl';
 import previewComponent from './SelectPreview';
-export const NetlifyCmsWidgetSelect = { controlComponent, previewComponent };
-export { controlComponent, previewComponent };
+
+const Widget = (opts = {}) => ({
+  name: 'select',
+  controlComponent,
+  previewComponent,
+  ...opts,
+});
+
+export const NetlifyCmsWidgetSelect = { Widget, controlComponent, previewComponent };
+export default NetlifyCmsWidgetSelect;