+++ packages/netlify-cms/src/editor-components.js
@@ -1,4 +1,0 @@
-import { registerEditorComponent } from 'netlify-cms-core';
-import image from 'netlify-cms-editor-component-image';
-
-registerEditorComponent(image);